requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.3
filelock==3.12.2
mysql-connector-python==8.1.0
platformdirs==3.10.0
protobuf==4.21.12
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.2
